+++ src/writers/csv.rs
@@ -2,7 +2,10 @@
 //! It presents the findings in a CSV format and prints it on STDOUT.
 
 use super::Writer;
-use crate::models::{reqres::UrlRequest, Finding};
+use crate::{
+    application::Args,
+    models::{reqres::UrlRequest, Finding},
+};
 
 /// A writer to print the findings as CSV.
 pub struct CsvWriter {
@@ -16,14 +19,15 @@ pub struct CsvWriter {
 
 impl Writer for CsvWriter {
     /// Create a new CsvWriter
-    fn new(ip_hostname: Option<String>, port: Option<u16>, url: Option<String>) -> Self {
-        let mut new_ip_hostname = ip_hostname;
-        let mut new_port = port;
+    fn new(argv: &Args) -> Self {
+        let mut new_ip_hostname = argv.ip_hostname.clone();
+        let mut new_port = argv.port;
+        let url = argv.url.clone();
 
         // If scan type is HTTP, we have the URL but no ip_hostname nor port.
         // define them from the URL
         if url.is_some() {
-            let url_request = UrlRequest::new(url.as_ref().unwrap(), false);
+            let url_request = UrlRequest::new(argv.url.as_ref().unwrap(), false);
             let (tmp_hostname, tmp_port) = url_request.get_hostname_port();
             new_ip_hostname = Some(tmp_hostname);
             new_port = Some(tmp_port);
@@ -32,7 +36,7 @@ impl Writer for CsvWriter {
         Self {
             ip_hostname: new_ip_hostname,
             port: new_port,
-            url,
+            url: url,
         }
     }